Abstract multtiple function calls
